backend/requirements.txt
Django==3.1.4
djangorestframework==3.9.0
gunicorn==20.0.4
psycopg2-binary==2.8.5